Fix bug: increase number of contributors per page (#2)
Fix a bug related to the GitHub API request for contributors: the number
of contributors returned in the JSON file is caped by the number of
entries per page, which is set by default to 30. Add an option to change
this value through the API and set its default to 1000.